wav-edges: Improve input loading, and print status info
This adds some informational printing of what the program is doing, and
how long each step took.

It also changes how the input waveform is loaded, by using the PCMBuffer
method instead of FullPCMBuffer; this makes it run faster and take less
memory (loading a 255M file went from needing 3.9G RAM to just 2.5G, and
took several seconds less time).

The decoder library seems to make some assumptions about reader behavior
that are not really guaranteed, but appear to work with the bytes.Reader
that we are currently using. The new code follows along, making similar
assumptions, since with the decoder API, it's tricky to do otherwise.