wav-edges: use go-arg for arg parsing, add param for stats
The printing of statistics is now off by default, but can be enabled via
a command-line argument.

Note: the latest release of go-arg does not show default values in the
help text for positional arguments (unlike other args), but the fix is
already in master, so will probably be included in the next release.
This commit works around that, with a TODO for cleanup after it updates.