dc-offset: add debug output, improve the filter a bit
The generated peak before the first peak is not as bad, and the sticky
offset after the last peak is no longer permanent, but both are still
much worse than I'd like to see.

This also simplifies the filter, looking at fewer peaks, as that turned
out to behave better.