dc-offset: allow in-place filtering, improve timings
This makes the filter output filtered samples rather than the offsets,
and restructures it to allow doing the filtering either in-place or to
a separate slice. This both makes it easier to use, and allows for less
memory use.

This also improves the CLI program's timings and stats a little bit.