pulse classifier: Peek ahead at lead-in instead of delaying
Previously, when the pulse classifier did not have a bit width, it would
calculate it from the second pulse it saw (the first after the leading
None), but would first return the first pulse (the leading None area) as
a PulseUnknown.

This changes it to instead peek ahead at the next 8 pulses on the first
call, averaging their widths to calculate the bit width immediately.
Thus it gets a better value for the bit width, and gets it sooner; it
now avoids returning PulseUnknown at all (unless this process fails).