pulse-classifier: Add an IIR filter on BitWidth adjustments
This seems to help, but feels like a hack - it slows down changes to the
bit width by only moving halfway from where it was to where the latest
pulse suggests it should be. Thus the old values never quite disappear.

Making it more heavily weighted towards the old values makes it seem to
work even better (with my data), but also feel even more like a hack.