Add interpolation of zero-crossings in edge detector
This outputs a float value in-between two samples, which should give
higher accuracy for when the waveform actually crossed zero.

In the case of edges from/to none, when no actual zero crossing was
found, the output float is just the unrounded value of the extrapolated
line's zero crossing point (to match the integer index).