pulse-stats: Fix first pair, reformat and reorder output
Previously, there was a spurious pulse pair at the start of the data,
counting the first pulse as a second pulse with nothing in front of it.
That has now been fixed to only consider actual pairs of pulses.

The output columns are now sized to the contents of each column, instead
of using one size for all of them, and also also reordered to group the
pulses instead of the mins/maxes/etc. (The rows are also reordered.)

This also changes the default output to be stdout instead of a file.